Split graphql db logic into multiple files. Add wrapper for dynamodb client. Add support for uuid to use as PK.
