+++ graphql/config/environment/development.js
@@ -1,3 +1,8 @@
 export default {
-  // Put your development configuration here
+	dynamoDB: {
+		region: 'us-west-2',
+	    endpoint: 'http://localhost:8000',
+	    accessKeyId: 'fakeKey123',
+	    secretAccessKey: 'fakeSecretKey456'
+	}
 };
