--- graphql/config/environment/test.js
@@ -1,3 +1,0 @@
-export default {
-  // Put your test configuration here
-};
