+++ graphql/database/tables.js
@@ -1,0 +1,5 @@
+const tableNames = {
+	features: 'Features',
+};
+
+module.exports = tableNames;
