+++ graphql/package.json
@@ -16,16 +16,19 @@
     "express-graphql": "^0.6.3",
     "graphql": "^0.9.1",
     "graphql-relay": "^0.5.1",
+    "uuid": "3.0.1",
     "webpack": "^2.2.1"
   },
   "devDependencies": {
     "aws-sdk": "^2.18.0",
     "babel-cli": "^6.22.2",
+    "jest": "^19.0.2",
     "require-clean": "^0.1.3"
   },
   "scripts": {
     "start": "babel-node ./server.js",
-    "update-schema": "babel-node ./scripts/updateSchema.js"
+    "update-schema": "babel-node ./scripts/updateSchema.js",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
